fix(gmail/parsers): extract shared extractVisibleText with safe entity decode (#453)
Resolves CodeQL js/double-escaping on payu.ts:27-35. The cascading
named-entity replace chain (with &amp; decoded last) could produce
double-decode paths like &amp;amp; → &amp; → &. The safe pattern is
a single /&([A-Za-z]+);/g regex with a lookup-table replacer so each
entity reference is matched exactly once and the result is never
re-scanned.

Extracted shared _text.ts with:
- NAMED_ENTITIES lookup covering the union of all four parsers
- decodeHtmlEntities (single-pass atomic, CodeQL-safe)
- extractVisibleText (strips style/script/comments/tags then decodes)

payu, wompi, mercado-pago, and apple now import from _text.ts.
bancolombia keeps its own implementation (different return type,
isolateTransactionalSentence coupling) — noted in _text.ts comment.
